Moved CommonComponent
