+++ src/components/Header/Header.js
@@ -1,4 +1,4 @@
-import { CommonComponent } from '@/common';
+import { CommonComponent } from '@/components/common';
 import { TagNames } from '@/utils';
 
 export class Header extends CommonComponent {
